Make Concept a class in it's own file
